Extract OVSP4RT logging utilities
- Extracted the FormatMac, LogErrorTable, and TableErrorMessage
  utility functions from es2k/ovsp4rt_es2k_public and moved them
  to common/ovsp4rt_logutils. Modified logging_test to use the
  common functions in place if its private copies.
