+++ ovs-p4rt/sidecar/common/CMakeLists.txt
@@ -5,6 +5,8 @@
 #
 
 target_sources(ovs_sidecar_o PUBLIC
+    ovsp4rt_logutils.cc
+    ovsp4rt_logutils.h
     ovsp4rt_private.cc
     ovsp4rt_private.h
     ovsp4rt_public.cc
