+++ ovs-p4rt/sidecar/es2k/ovsp4rt_es2k_public.cc
@@ -5,7 +5,7 @@
 //
 
 #include "absl/flags/flag.h"
-#include "absl/strings/str_cat.h"
+#include "common/ovsp4rt_logutils.h"
 #include "common/ovsp4rt_private.h"
 #include "common/ovsp4rt_utils.h"
 #include "lib/ovsp4rt_credentials.h"
@@ -14,29 +14,6 @@
 #include "openvswitch/ovs-p4rt.h"
 #include "ovsp4rt_es2k_defs.h"
 #include "ovsp4rt_es2k_private.h"
-
-namespace {
-
-const std::string TableErrorMessage(bool inserting, const char* table) {
-  if (inserting) {
-    return absl::StrCat("Error adding entry to ", table);
-  } else {
-    return absl::StrCat("Error deleting entry from ", table);
-  }
-}
-
-const std::string FormatMac(const uint8_t* mac_addr) {
-  char buf[32];
-  snprintf(buf, sizeof(buf), "%x:%x:%x:%x:%x:%x", mac_addr[0], mac_addr[1],
-           mac_addr[2], mac_addr[3], mac_addr[4], mac_addr[5]);
-  return buf;
-}
-
-void LogTableError(bool inserting, const char* table) {
-  ovsp4rt_log_error("%s", TableErrorMessage(inserting, table).c_str());
-}
-
-}  // namespace
 
 //----------------------------------------------------------------------
 // Functions with C interfaces
