+++ ovs-p4rt/sidecar/testing/logging_test.cc
@@ -1,3 +1,6 @@
+// Copyright 2024 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
 #define _POSIX_SOURCE
 #include <limits.h>
 #include <stdbool.h>
@@ -8,28 +11,11 @@
 #include <iostream>
 #include <string>
 
-#include "absl/strings/str_cat.h"
+#include "common/ovsp4rt_logutils.h"
 #include "lib/ovsp4rt_logging.h"
 #include "openvswitch/ovs-p4rt.h"
 
-const std::string TableErrorMessage(bool inserting, const char* table) {
-  if (inserting) {
-    return absl::StrCat("Error adding entry to ", table);
-  } else {
-    return absl::StrCat("Error deleting entry from ", table);
-  }
-}
-
-const std::string FormatMac(const uint8_t* mac_addr) {
-  char buf[32];
-  snprintf(buf, sizeof(buf), "%x:%x:%x:%x:%x:%x", mac_addr[0], mac_addr[1],
-           mac_addr[2], mac_addr[3], mac_addr[4], mac_addr[5]);
-  return buf;
-}
-
-void LogTableError(bool inserting, const char* table) {
-  ovsp4rt_log_error("%s", TableErrorMessage(inserting, table).c_str());
-}
+using namespace ovs_p4rt;
 
 void log_messages() {
   uint8_t mac_addr[6] = {0xb, 0xe, 0xe, 0xb, 0xe, 0xe};
